Create a data.tree object to be used for aggregating or scaling interval data

Create a data.tree object to be used for aggregating or scaling interval data

create_agg_interval_tree(data_intervals_dt, agg_intervals_dt, col_stem)

create_scale_interval_tree(data_intervals_dt, col_stem)

Arguments

data_intervals_dt
[data.table()]
Describes each interval present in the data. Includes a column for '{col_stem}_start', '{col_stem}_end', and col_stem.

agg_intervals_dt
[data.table()]
Describes each interval that needs to be aggregated to. Includes a column for '{col_stem}_start', '{col_stem}_end', and col_stem.

col_stem
[character(1)]
The name of the variable to be aggregated or scaled over. If aggregating an 'interval' variable should not include the '_start' or '_end' suffix.

Value

[data.tree()] where the 'name' field of each node is in interval notation and describes each left-closed, right-open interval. Each node also includes a field for the 'left' and 'right' endpoint.

Details

create_agg_interval_tree() returns a data.tree() with three levels. The root node is a place holder that covers the entire data interval. The second level contains each aggregate node that needs to be made and each of these aggregate node's have children nodes for the data intervals that are needed for aggregation.

create_scale_interval_tree() returns a data.tree() with a variable number of levels dependent on the intervals available in the data. The root node also covers the entire data interval, this node may or may not actually be in data_intervals_dt. Each interval in data_intervals_dt is then positioned in the tree so that it is a sub-interval of its parent interval node.