Identify issues in the input dataset for aggregation and scaling functions

Identify issues in the input dataset for aggregation and scaling functions

identify_agg_dt_issues(
  dt,
  id_cols,
  value_cols,
  col_stem,
  col_type,
  mapping,
  agg_function = sum,
  drop_present_aggs = FALSE,
  collapse_interval_cols = FALSE
)

identify_scale_dt_issues(
  dt,
  id_cols,
  value_cols,
  col_stem,
  col_type,
  mapping,
  agg_function = sum,
  collapse_missing = FALSE,
  collapse_interval_cols = FALSE
)

Arguments

dt
[data.table()]
Data to be aggregated or scaled.

id_cols
[character()]
ID columns that uniquely identify each row of dt.

value_cols
[character()]
Value columns that should be aggregated.

col_stem
[character(1)]
The name of the variable to be aggregated or scaled over. If aggregating an 'interval' variable should not include the '_start' or '_end' suffix.

col_type
[character(1)]
The type of variable that is being aggregated or scaled over. Can be either 'categorical' or 'interval'.

mapping
[data.table()]
For 'categorical' variables, defines how different levels of the hierarchical variable relate to each other. For aggregating 'interval' variables, it is used to specify intervals to aggregate to, while when scaling the mapping is inferred from the available intervals in dt.

agg_function
[function()]
Function to use when aggregating, can be either sum (for counts) or prod (for probabilities).

drop_present_aggs
[logical(1)]
Whether to drop aggregates (or overlapping intervals) that are already present in dt before aggregating. Default is 'False' and the function errors out.

collapse_interval_cols
[logical(1)]
Whether to collapse interval id_cols (not including col_stem if it is an interval variable). Default is 'False'. If set to 'True' the interval columns are collapsed to the most detailed common intervals and will error out if there are overlapping intervals. See details or vignettes for more information.

collapse_missing
[logical(1)]
When scaling a categorical variable, whether to collapse missing intermediate levels in mapping. Default is 'False' and the function errors out due to missing data.

Value

[data.table()] with problematic rows in dt. Only includes original id_cols columns and a new column "issue" describing each row's issue.

Details

identify_agg_dt_issues() steps:

For interval id variables (not including col_stem) if collapse_interval_cols = TRUE, checks for overlapping intervals and missing intervals and then collapses these interval variables to the most detailed common intervals. 2 If col_type = "interval" then also checks for overlapping intervals and missing intervals but doesn't collapse this variable.

Creates data.tree from mappings to be used for aggregation.

If col_type = "categorical" checks for aggregates already present.

Identifies any missing nodes in data.tree needed to make aggregations.

identify_scale_dt_issues() steps:

For interval id variables (not including col_stem) if collapse_interval_cols = TRUE, checks for overlapping intervals and missing intervals and then collapses these interval variables to the most detailed common intervals.

Creates data.tree from mappings to be used for scaling.

Identifies any missing nodes in data.tree needed to make aggregations.

Examples

## Try to aggregate data with a location missing errors out
input_dt <- data.table::CJ(location = iran_mapping[!grepl("[0-9]+", child),
                                                   child],
                           year = 2011, value = 1)
input_dt <- input_dt[location != "Tehran"]
if (FALSE) {
output_dt <- agg(dt = input_dt,
                 id_cols = c("location", "year"),
                 value_cols = "value",
                 col_stem = "location",
                 col_type = "categorical",
                 mapping = iran_mapping)
}
issues_dt <- identify_agg_dt_issues(dt = input_dt,
                                    id_cols = c("location", "year"),
                                    value_cols = "value",
                                    col_stem = "location",
                                    col_type = "categorical",
                                    mapping = iran_mapping)

## Try to scale data with a location missing errors out
input_dt <- data.table::CJ(location = iran_mapping[!grepl("[0-9]+", child),
                                                   child],
                           year = 2011,
                           value = 1)
input_dt_agg <- data.table::data.table(
  location = "Iran (Islamic Republic of)",
  year = 2011, value = 62
)
input_dt <- rbind(input_dt, input_dt_agg)
if (FALSE) {
output_dt <- scale(dt = input_dt,
                   id_cols = c("location", "year"),
                   value_cols = "value",
                   col_stem = "location",
                   col_type = "categorical",
                   mapping = iran_mapping,
                   collapse_missing = TRUE)
}
issues_dt <- identify_scale_dt_issues(dt = input_dt,
                                      id_cols = c("location", "year"),
                                      value_cols = "value",
                                      col_stem = "location",
                                      col_type = "categorical",
                                      mapping = iran_mapping,
                                      collapse_missing = TRUE)