Is the interval variable missing any expected intervals?

Checks to see if the input interval variable is missing any expected intervals.

assert_no_missing_intervals(ints_dt, expected_ints_dt)

identify_missing_intervals(ints_dt, expected_ints_dt)

Arguments

ints_dt
[data.table()]
Unique intervals. The first column represents the start of each interval and the second column represents the end of each interval.

expected_ints_dt
[data.table()]
The expected intervals that should be completely included in ints_dt. The first column represents the start of each interval and the second column represents the end of each interval.

Value

identify_missing_intervals returns a [data.table()] with columns for the 'start' and 'end' of the missing intervals. If no intervals are missing then a zero-row [data.table()] is returned. assert_no_missing_intervals returns nothing but throws an error if identify_missing_intervals returns a non-empty data.table.

Examples

ints_dt <- data.table::data.table(
  start = seq(0, 95, 5),
  end = c(seq(5, 95, 5), Inf)
)
missing_dt <- identify_missing_intervals(
  ints_dt = ints_dt[!start %in% c(0, 10, 95)],
  expected_ints_dt = data.table::data.table(start = 0, end = Inf)
)