Does the interval variable have any overlapping intervals?

Checks to see if the input interval variable has any overlapping intervals.

assert_no_overlapping_intervals(ints_dt)

identify_overlapping_intervals(ints_dt)

Arguments

ints_dt
[data.table()]
Unique intervals. The first column represents the start of each interval and the second column represents the end of each interval.

Value

identify_overlapping_intervals returns a [data.table()] with columns for the 'start' and 'end' of the overlapping intervals. If no intervals are overlapping then a zero-row [data.table()] is returned. assert_no_overlapping_intervals returns nothing but throws an error if identify_overlapping_intervals returns a non-empty data.table.

Examples

ints_dt <- data.table::data.table(
  start = c(seq(0, 95, 5), 0),
  end = c(seq(5, 95, 5), Inf, Inf)
)
overlapping_dt <- identify_overlapping_intervals(ints_dt)