Subset to data for one unique set of the agg/scale variable

Subset to data for one unique set of the agg/scale variable

subset_unique_grouping(
  dt,
  id_cols,
  col_stem,
  col_type,
  collapse_interval_cols,
  groups,
  group_num
)

Arguments

dt
[data.table()]
Data to be aggregated or scaled.

id_cols
[character()]
ID columns that uniquely identify each row of dt.

col_stem
[character(1)]
The name of the variable to be aggregated or scaled over. If aggregating an 'interval' variable should not include the '_start' or '_end' suffix.

col_type
[character(1)]
The type of variable that is being aggregated or scaled over. Can be either 'categorical' or 'interval'.

collapse_interval_cols
[logical(1)]
Whether to collapse interval id_cols (not including col_stem if it is an interval variable). Default is 'False'. If set to 'True' the interval columns are collapsed to the most detailed common intervals and will error out if there are overlapping intervals. See details or vignettes for more information.

groups
[list(2)]
Output of identify_unique_groupings().

group_num
[integer(1)]
Integer row of groups$unique_groupings to subset data to.

Value

list with two named elements.

unique_cols: [data.table()] with unique combinations of the variable being aggregated/scaled. If col_type is 'categorical' then just a column for col_stem. If col_type is interval' then a column for '{col_stem}_start' and '{col_stem}_end' then the 'start' and 'end' variable values are also ',' separated.

dt: [data.table()] subset of the input dt that has a specific combination of the variable being aggregated or scaled over.

See also